+++ pycmds/_main_window.py
@@ -195,7 +195,6 @@ class MainWindow(QtWidgets.QMainWindow):
         from .hardware.spectrometers import spectrometers
         from .hardware import delays
         from .hardware.filters import filters
-        import pycmds._record
         import pycmds._sensors
 
     def _initialize_widgets(self):
